requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
bcrypt==4.0.1
billiard==4.2.0
bitarray==2.9.2
celery==5.3.6
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cytoolz==0.12.3
dj-database-url==1.0.0
Django==4.2.11
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-heroku==0.3.1
django-postgres-extensions==0.9.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
frozenlist==1.4.1
gunicorn==20.1.0
hexbytes==0.3.1
idna==3.4
jsonpickle==2.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
lru-dict==1.3.0
multidict==6.0.5
parsimonious==0.9.0
prompt-toolkit==3.0.43
protobuf==5.26.0
psycopg2==2.9.9
pycryptodome==3.20.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.6
pytz==2022.5
pyunormalize==15.1.0
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.28.1
rlp==4.0.0
rpds-py==0.18.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.3
toolz==0.12.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.12
uWSGI==2.0.24
vine==5.1.0
w3lib==2.0.1
wcwidth==0.2.13
web3==6.11.2
websockets==12.0
whitenoise==6.6.0
yarl==1.9.4
